refactor: internal refactor to simplify template names
