+++ lib/module.js
@@ -22,8 +22,8 @@ module.exports = function (moduleOptions) {
 
   // Plugin
   const { dst } = this.addTemplate({
-    src: resolve(__dirname, './templates/auth.plugin.js'),
-    fileName: 'auth/auth.plugin.js',
+    src: resolve(__dirname, './templates/plugin.js'),
+    fileName: 'auth/plugin.js',
     options
   })
 
@@ -33,22 +33,22 @@ module.exports = function (moduleOptions) {
 
   // Middleware
   this.addTemplate({
-    src: resolve(__dirname, './templates/auth.middleware.js'),
-    fileName: 'auth/auth.middleware.js',
+    src: resolve(__dirname, './templates/middleware.js'),
+    fileName: 'auth/middleware.js',
     options
   })
 
   // Utilities
   this.addTemplate({
-    src: resolve(__dirname, './templates/auth.utilities.js'),
-    fileName: 'auth/auth.utilities.js',
+    src: resolve(__dirname, './templates/utilities.js'),
+    fileName: 'auth/utilities.js',
     options
   })
 
   // Class
   this.addTemplate({
-    src: resolve(__dirname, './templates/auth.class.js'),
-    fileName: 'auth/auth.class.js',
+    src: resolve(__dirname, './templates/auth.js'),
+    fileName: 'auth/auth.js',
     options
   })
 }
