+++ lib/templates/auth.js
@@ -1,8 +1,8 @@
 import Cookies from 'js-cookie'
 import { parse as parseCookie } from 'cookie'
-import { routeOption } from './auth.utilities'
 import getProp from 'dotprop'
 import Vue from 'vue'
+import { routeOption } from './utilities'
 
 const RELATIVE_URL_REGEX = /^\/[a-zA-Z0-9@\-%_~][\/a-zA-Z0-9@\-%_~]{1,200}$/
