+++ lib/templates/middleware.js
@@ -1,5 +1,5 @@
 import Middleware from '../middleware'
-import { routeOption } from "./auth.utilities";
+import { routeOption } from "./utilities"
 
 Middleware.auth = function (ctx) {
   // Disable middleware if options: { auth: false } is set on the route
