+++ lib/templates/plugin.js
@@ -1,6 +1,5 @@
-import './auth.middleware'
-
-import Auth from './auth.class'
+import Auth from './auth'
+import './middleware'
 
 export default function (ctx, inject) {
   // Create  new Auth instance
